Add support to pass option on build
Options can be passed using `--build-arg` with `docker build` and
`--opt` on `buildctl`. As of today, the only supported option is to
enable tekton oci bundle configuration, but we can think of other
options in the futures (override parameters, pass runtime options, …)

It also add the enable-tekton-oci-bundles options to true on our test
for now. Long term, we need a better set of tests suites.

This also clean some logs.

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>